pyproject.toml
[build-system]
requires = [
  "setuptools>=45", 
  "wheel", 
  "setuptools_scm[toml]>=6.2",
]
build-backend = "setuptools.build_meta"

[project]
name = "policyengine_us_data"
version = "1.63.0"
description = "A package to create representative microdata for the US."
readme = "README.md"
authors = [
    {name = "PolicyEngine", email = "hello@policyengine.org"},
]
license = {file = "LICENSE"}
requires-python = ">=3.12, <3.14.0"
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "policyengine-us>=1.516.0",
    "policyengine-core>=3.23.6",
    "pandas>=2.3.1",
    "requests>=2.25.0",
    "tqdm>=4.60.0",
    "microdf_python>=1.2.1",
    "setuptools>=60",
    "microimpute>=1.1.4",
    "pip-system-certs>=3.0",
    "google-cloud-storage>=2.0.0",
    "google-auth>=2.0.0",
    "scipy>=1.15.3",
    "statsmodels>=0.14.5",
    "openpyxl>=3.1.5",
    "tables>=3.10.2",
    "torch>=2.7.1",
    "us>=2.0.0",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
    "xlrd>=2.0.2",
    "spm-calculator>=0.1.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
calibration = [
    "samplics",
]
l0 = [
    "l0-python",
]

[dependency-groups]
dev = [
    "black",
    "pytest",
    "quantile-forest",
    "tabulate",
    "furo",
    "jupyter-book",
    "mystmd>=1.7.0",
    "yaml-changelog>=0.1.7",
    "build",
    "tomli",
    "itables",
]

[tool.setuptools]
packages = ["policyengine_us_data"]
include-package-data = true

[tool.setuptools.package-data]
"policyengine_us_data" = ["**/*"]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = [
    "policyengine_us_data/tests",
]

[tool.black]
line-length = 79
target-version = ['py311', 'py312', 'py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''
